refactor(ar-analytics): extract 4 tab panels, no visual change
Pure move ahead of the restyle so move bugs and style bugs stay separable.
Shared types, aging/channel colours, and formatters go to ar/shared.tsx.
